+++ docs/CHANGELOG.md
@@ -57,6 +57,7 @@
 - changed animated spikes to work outside levels 5 and 7
 - changed `O_AI_PATROL_1` behavior in High Security Compound and Area 51 to be configurable; refer to migration docs
 - removed the hard-coded Aldwych drill speed override and moved it to Lua instead
+- fixed handheld flare sparks appearing inside the flare instead of above the tip
 - fixed TR3:LA referring to a non-existing FMV in the startup logs
 - fixed texture bleeding and missized textures on piranhas, tropical fish and bats
 - fixed the drill in the Shakespeare Cliff not rotating